test_requirements.txt
pytest
pytest-pep8
pytest-cov
django
djangorestframework
